pyproject.toml
[project]
name = "dataflowhub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "dj-database-url>=3.0.1",
    "django>=5.2.8",
    "django-filter>=25.2",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "flake8>=7.3.0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "redis>=7.0.1",
    "requests>=2.32.5",
]
